Rename chat-contents.ts to chat-content.ts
