+++ backend/src/app.ts
@@ -11,7 +11,7 @@ import apiKey from "./routes/private/api-key";
 import privateGpis from "./routes/private/gpis";
 import publicGpis from "./routes/public/gpis";
 import chat from "./routes/public/chat";
-import chatContents from "./routes/private/chat-contents";
+import chatContent from "./routes/private/chat-content";
 import histories from "./routes/private/histories";
 import providerTypes from "./routes/public/provider-types";
 import session from "./routes/public/session";
@@ -45,7 +45,7 @@ fastify.register(prismaPlugin);
 fastify.get("/health", () => true);
 fastify.register(apiKey, { prefix: "/api/keys" });
 fastify.register(chat, { prefix: "/chat" });
-fastify.register(chatContents, { prefix: "/chat/contents" });
+fastify.register(chatContent, { prefix: "/chat/contents" });
 fastify.register(privateGpis, { prefix: "/users/gpis" });
 fastify.register(publicGpis, { prefix: "/gpis" });
 fastify.register(histories, { prefix: "/histories" });
