+++ backend/src/routes/private/chat-content.ts
